requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
groq==0.4.1
pydantic==2.5.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch
sentence-transformers==2.2.2
scikit-learn==1.3.2
